+++ Chromatics/Layers/LayerProcessorFactory.cs
@@ -80,6 +80,8 @@ namespace Chromatics.Layers
                 EffectLayerType.DamageFlash => DamageFlashProcessor.Instance,
                 EffectLayerType.GoldSaucerVegas => GoldSaucerVegasProcessor.Instance,
                 EffectLayerType.CutsceneAnimation => CutsceneAnimationProcessor.Instance,
+                EffectLayerType.CastingSuccess => CastingSuccessProcessor.Instance,
+                EffectLayerType.StatusInflicted => StatusInflictedProcessor.Instance,
                 _ => throw new ArgumentException("Unknown EffectLayerType")
             };
         }