+++ Chromatics/Models/EffectTypesModel.cs
@@ -18,6 +18,8 @@ namespace Chromatics.Models
         public bool effect_startupanimation { get; set; } = true;
         public bool effect_raideffects { get; set; } = true;
         public bool effect_audiovisualizer { get; set; } = false;
+        public bool effect_castingsuccess { get; set; } = true;
+        public bool effect_statusinflicted { get; set; } = false;
 
         //Specific Effect Settings
         public bool effect_damageflash_scaledamage { get; set; } = true;